# shared/openapi/mobile-api-v1.yaml
openapi: 3.1.0
info:
  title: Khoya Paya Mobile API
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
paths:
  /auth/login:
    post:
      summary: Create a device-bound mobile session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "201": { description: Token pair and safe user DTO }
        "400": { $ref: "#/components/responses/ApiError" }
        "401": { $ref: "#/components/responses/ApiError" }
        "403": { $ref: "#/components/responses/ApiError" }
        "429": { $ref: "#/components/responses/ApiError" }
...
components:
  securitySchemes:
    mobileBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Device:
      type: object
      required: [deviceId, platform]
      properties:
        deviceId: { type: string, minLength: 8, maxLength: 200 }
        platform: { type: string, enum: [android, ios] }
        deviceName: { type: string, maxLength: 120 }
        appVersion: { type: string, maxLength: 40 }
    LoginRequest:
      type: object
      required: [identifier, password, device]
      properties:
        identifier: { type: string, format: email }
        password: { type: string, format: password }
        device: { $ref: "#/components/schemas/Device" }
    ApiError:
      type: object
      required: [success, error]
      properties:
        success: { const: false }
        error:
          type: object
          required: [code, message, fieldErrors]
          properties:
            code:
              enum: [VALIDATION_ERROR, INVALID_CREDENTIALS, AUTHENTICATION_REQUIRED, TOKEN_EXPIRED, REFRESH_TOKEN_INVALID, SESSION_REVOKED, ACCOUNT_BLOCKED, ACCOUNT_SUSPENDED, FORBIDDEN, RATE_LIMITED, NOT_FOUND, INTERNAL_ERROR]
            message: { type: string }
            fieldErrors: { type: object, additionalProperties: { type: array, items: { type: string } } }
  responses:
    ApiError:
      description: Stable mobile API error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }